Refactor academic controllers to use updated validation schemas and optimize admin updates

+++ src/controller/academics/academicYear.controller.js
@@ -6,16 +6,16 @@ import Admin from '../../model/Staff/Admin.model.js';
 import {
 	createAcademicYearSchema,
 	updateAcademicYearSchema,
-} from '../../schemas/academicYear.schema.js';
+
+} from '../../utils/Validation/academicYear.schema.js';
+
 
 //*	@ desc Create a new academic year
 //*	@ route POST /api/v1/academics/academic-year
 //*	@ access Private
 
 export const createAcademicYear = AsyncHandler(async (req, res) => {
-	const { name, fromYear, toYear } = createAcademicYearSchema.parse(
-		req.body
-	);
+	const { fromYear, toYear, name } = createAcademicYearSchema.parse(req.body);
 	const createdBy = req.user._id;
 
 	const academicYear = await AcademicYear.findOne({ name });
@@ -32,10 +32,13 @@ export const createAcademicYear = AsyncHandler(async (req, res) => {
 	});
 
 	// push the academic year to the user
-	const admin = await Admin.findById(req.user._id);
-	admin.academicYears.push(academicYearCreated._id);
+	// const admin = await Admin.findById(req.user._id);
+	// admin.academicYears.push(academicYearCreated._id);
 
-	await admin.save();
+	await Admin.updateOne(
+		{ _id: req.user._id },
+		{ $push: { academicYears: academicYearCreated._id } }
+	)
 
 	res.status(201).json(
 		new ApiResponse(
@@ -53,10 +56,13 @@ export const getAcademicYears = AsyncHandler(async (req, res) => {
 	const academicYears = await AcademicYear.find({}).sort({
 		createdAt: 1,
 	});
+
+	const count = await AcademicYear.countDocuments();
+
 	res.status(200).json(
 		new ApiResponse(
 			'Academic years fetched successfully',
-			academicYears
+			{ academicYears, total: count }
 		)
 	);
 });
@@ -66,6 +72,10 @@ export const getAcademicYears = AsyncHandler(async (req, res) => {
 //* @ access Private
 
 export const getAcademicYearById = AsyncHandler(async (req, res) => {
+
+	if (!req.params.id) {
+		throw new ApiError(400, 'Please provide an id');
+	}
 	const academicYear = await AcademicYear.findById({
 		_id: req.params.id,
 	});
@@ -87,15 +97,27 @@ export const getAcademicYearById = AsyncHandler(async (req, res) => {
 //* @ access Private
 
 export const updateAcademicYear = AsyncHandler(async (req, res) => {
-	const { name, fromYear, toYear } = updateAcademicYearSchema.parse(
-		req.body
-	);
-	const updatedBy = req.user._id;
+	if (!req.params.id) {
+		throw new ApiError(400, 'Please provide an id');
+	}
+	let { fromYear, toYear, name } = updateAcademicYearSchema.parse(req.body);
+
 	const AcademicYearExist = await AcademicYear.findById(
 		req.params.id
 	);
 	if (!AcademicYearExist) {
 		throw new ApiError(404, 'Academic year not found');
+	}
+
+	const updatedBy = req.user._id;
+	name = name !== undefined
+		? name
+		: `${fromYear !== undefined ? fromYear : AcademicYearExist.fromYear}-${toYear !== undefined ? toYear : AcademicYearExist.toYear}`;
+	fromYear = fromYear !== undefined ? fromYear : AcademicYearExist.fromYear
+	toYear = toYear !== undefined ? toYear : AcademicYearExist.toYear
+
+	if (fromYear >= toYear) {
+		throw new ApiError(400, 'The first year must be less than the second year');
 	}
 
 	const academicYear = await AcademicYear.findOne({ name });
@@ -104,15 +126,21 @@ export const updateAcademicYear = AsyncHandler(async (req, res) => {
 		throw new ApiError(400, 'Academic year already exist');
 	}
 
+	const updateAcademicYearQuery = {
+		name,
+		fromYear,
+		toYear,
+		updatedBy,
+	};
+
+
 	const academicYearUpdated = await AcademicYear.findByIdAndUpdate(
 		req.params.id,
+		updateAcademicYearQuery,
 		{
-			name,
-			fromYear,
-			toYear,
-			updatedBy,
-		},
-		{ new: true }
+			new: true,
+			runValidators: true,
+		}
 	);
 
 	if (!academicYearUpdated) {
@@ -132,6 +160,9 @@ export const updateAcademicYear = AsyncHandler(async (req, res) => {
 //* @ access Private
 
 export const deleteAcademicYear = AsyncHandler(async (req, res) => {
+	if (!req.params.id) {
+		throw new ApiError(400, 'Please provide an id');
+	}
 	const academicYear = await AcademicYear.findByIdAndDelete(
 		req.params.id
 	).select('name -_id');
